requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.3
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
psycopg==3.1.8
psycopg2-binary==2.9.3
google-generativeai==0.8.3
bs4==0.0.2
requests==2.32.3
django_qr_code==4.1.0
pillow==11.1.0
google==3.0.0
google-genai==0.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
setuptools==76.0.0
drf-spectacular==0.28.0